Added clean_tf() function
Added the clean_tf() function at the bottom of xferfcn.py, adjusting it slightly to be conform with the existing imports in the file and adding one import from another file of the control library. Also adjusted the documentation for the function to match the existing style. Additionally fixed a formatting error in the documentation for clean_ss()