Added fival() function
Added the fival() function at the bottom of thetimeresp.py file, slightly adjusting it to conform with the existing imports in the file and adding a couple of new import statements. Adjusted the documentation so that its format is conform with the rest of the file. Slightly adjusted the format of the documentation for clean_tf() in the xferfcn.py file