Added the clean() and fival() functions as functions to the TransferFunction and StateSpace classes
While the functions clean_ss and clean_tf remain as separate functions in the respective files, the function clean() has been added in the respective form to the TransferFunction ans StateSpace classes. This makes it possible to still go directly from a StateSpace to a clean TransferFunction (or vice versa) but makes it easier to clean either one of the two if it is already in the desired form. The fival function remains as a function in the timeresp.py file, but by importing it to the other two files and including it in the aforementioned classes it is now possible to, for instance, use sys.fival(**) where sys is a StateSpace or TransferFunction.